Implement measurement rollback on Android S
Since we do not have access to System Server on S, we instead store data in AppSearch. Also, we don't have the necessary MANAGE_ROLLBACKS permission to use `RollbackManager` to query recent mainline train rollbacks. Therefore, we instead store the current apex version in a document in AppSearch when a deletion event occurs, and compare this stored value to the current apex version on Measurement service start. If the stored value is higher, then it indicates that a deletion event happened in a newer apex version, which means that a rollback happened and the measurement db needs to be cleared.

Bug: 280727787
Test: atest AdExtServicesServiceCoreUnitTests
Test: atest AdServicesEndToEndTests
Test: atest CtsAdServicesEndToEndTestMeasurement

Change-Id: Ic2a608508dab99a9103ce007076ca750f14facc4