Parse DeviceConfig content provider allowlist
A proto format is defined for a content provider allowlist. The
allowlist consists of a list of content provider authoritiy strings.

The allowlist can be pushed via DeviceConfig (feature flags will be
added later). This will be in the form of a base64 encoded string, which
will need to be parsed and added to the existing allowlist in
SdkSandboxManagerService.

Bug: 274070295
Test: atest
ContentProviderRestrictionsHostTest#testGetContentProvider_DeviceConfigAllowlistApplies

Change-Id: I98c3b2756e3a95198018f06f40dfa6231b28e32f