Invoke death callback once binding has died
If we call death callback in onSdkSandboxDied and the app decides to
call loadSdk again, then there is a possible race between the binding of
the new sandbox and the unbinding of the old sandbox, which leads to
onConnected never being called.

Instead, we now invoke the death callback once the old sandbox
binding dies and when a sandbox dies which was already unbound.
This fixes the flakiness of SdkSandboxMetricsHostTest#testSdkCanAccessSdkSandboxExitReasons
, which immediately calls loadSdk when the sandbox dies.

Test: atest com.android.tests.sdksandbox.host.SdkSandboxMetricsHostTest --iterations 15
Bug: 282666880

Change-Id: If4e6e3afd7950dae746d946a9b5bc2ae11807b68