Initial usage of AdServicesExtendedMockitoRule.
It's replacing the usages of StaticMockFixtureRule (all occurrences)
and mockito session management on @Before / @After (for now just on
DialogFragmentTest, but other tests will used it as well).

Notice that the main logic is defined by ExtendedMockitoRule, which
will be provided by module-utils but it's not available in this
branch yet (hence it's copy-and-pasted into this project).

Test: atest \
  AdServicesServiceCoreUnitTests:E2EDenoisedMockTest \
  AdServicesServiceCoreUnitTests:E2EImpressionNoiseMockTest \
  AdServicesServiceCoreUnitTests:E2EInteropMockTest \
  AdServicesServiceCoreUnitTests:com.android.adservices.service.PhFlagsTest \
  AdServicesServiceCoreUnitTests:SysPropForceDefaultValueFixtureTest

Bug: 281577492
Bug: 280899479

Change-Id: I342ae3b37194862323b4b7001188defedd4560b6