Change frequency cap to a max rate of occurrences
Previously, frequency caps were used to specify the maximum number of
previous occurrences of an event.  It makes more intuitive sense to
instead use frequency caps as a maximum rate of occurrence.

So instead of maxCount = 2, interval = 1 day meaning "show this ad if it
hasn't been shown two times in the past day already" the frequency cap
now reads "occurrences of this event should not happen more than twice a
day."

Test: atest AdServicesServiceCoreUnitTests AdServicesFrameworkUnitTests
Fixes: 278275929
Change-Id: I9180c41b95208078f15fb4085dde6526215fcee1