Address flaky expiration time in test
When a custom audience is joined, it is validated to ensure that no bad
data causes the service to behave unexpectedly.  One of these
validations is that the expiration time must be in the future (from the
instant that validation occurs).

On lower-performance devices, an expiry that is not too far in the
future could cause the test to fail because the device took too long
to process a joinCustomAudience request.  To address this, the CA with
early expiry will be joined first.

Also added more logging to the Custom Audience APIs since validation
errors are lost.

Test: atest CtsAdServicesDebuggableDeviceTestCases
Bug: 283915281
Change-Id: I0dc643bd75e43ba8f3002d2099ebae90928ad00a