Fix background job crash on static init due to missing LoggerFactory
When a background job is scheduled persistently on S, and then the device OTAs to T+, the background job still tries to execute on T. We have code in the onStartJob method that detects this condition and terminates the job. However, some jobs (like EpochJob) use the LoggerFactory class in the static initializer. This LoggerFactory class was only added to AdServices in the M05 train. Therefore, if the device OTAs to an AdServices build that's before 05, the job crashes with a class-not-found exception. Resolving this by removing dependency on LoggerFactory in the static initializer of the background jobs.

Bug: 280993428
Test: atest AdExtServicesServiceCoreUnitTests
Test: atest AdServicesServiceCoreUnitTests
Change-Id: I4bff6a62d439e1a2effcfdc1c15462b4ab391415