Make FrequencyCapFilters a collection of lists
A Set<KeyedFrequencyCap> is not a full promise or guarantee that the
frequency caps added will make sense; namely, two caps can be added with
the same key and interval which only differ in the max count.

Instead, don't make any guarantees about the collection of frequency
caps and just deduplicate internally.

Test: atest AdServicesServiceCoreUnitTests AdServicesFrameworkUnitTests
Bug: 285363029
Change-Id: I36fc7e0b3ed7b2a48232598ef9a4c57a3c0de3ef