[fetchCA] Add CustomAudienceBlob
A CustomAudienceBlob is a generalized representation of a custom
audience which can be constructed to/from any of its use-case specific
alternatives (FetchCustomAudienceInput, JSONObject,
DBCustomAudience and more). This makes testing and development of the
fetchAndJoinCustomAudience API more feasible and simple. It can also
simplify other existing APIs as part of future cleanups.

Test: atest AdServicesServiceCoreUnitTests

Bug: 283857101

Change-Id: I9834b0e4ee2d4117b5f3cff520796b9d80082ae5