Parse DeviceConfig content provider canary set allowlist
A proto format is defined for a content provider canary allowlist. This
allowlist consists of a list of content provider authority strings.

This allowlist will be pushed via DeviceConfig in the form of a base64
encoded string, which will need to be parsed and added to the existing
allowlist in SdkSandboxManagerService.

Bug: 274070295
Test: atest ContentProviderRestrictionsHostTest
Change-Id: I76bed16b13d56ef783b2ec552cc276cffe2e08f3