CTS - SDK can disable and enable sandbox activity back navigation
This change returns new Binder (ActivityActionExecutor) as a way for the
test to send signals to the test SDK to apply changes over the sandbox
activity. It comes with 2 functionalities to enable and disable back
navigation.

That is tested using the UiDevice::pressBack() button which should work
only if the back navigation is enabled.

Test: atest SdkSandboxManagerTest
Bug: 288892796
Change-Id: I59266493f41d5e43ff384eacb508a0d5567d4d38