Include Cobalt registry in `service-core`
The Cobalt registry is checked in as a text proto and is converted to a
binary proto at build time with a custom genrule. `service-core` then
includes the generated binary proto as a Java resource and parses into
an object at runtime.

This is the same approach used to embed the list of platform APIs
allowed per target SDK in
//packages/modules/AdServices/sdksandbox/service/Android.bp

Test: atest CobaltRegistryLoaderTest

Change-Id: I066d3098f37d1c288ff165c0d0861d2dfd3a95c3