Change KeepSandboxAliveRule to load test sdk
Loading and unloading test sdk for each test is causing test crashes on
low-memory devices because SdkSandboxManager.unlaodSdk does not
guarantee that memory associated with the sdk is freed (only the
references are dropped).

Modifying this class rule to load the test sdk at the beginning of the
test suite and unloading it at the end. Before each test,
SdkSandboxScenarioRule checks if an sdk is already loaded through the
SdkSandboxManager, and if not, it attempts to load the sdk itself.

ag/22692795 was reverted due to b/279629307. Fixing it back.

Bug: b/274296116
Test: atest CtsSdkSandboxWebkitTestCases

Change-Id: Ie76a6040e7171b748cb6a7300182fa17481765aa