Fix unit test that's failing due to code still running from preceding tests
These tests include a verifyNoMoreInteractions(staticMockMarker(TopicsWorker.class)) in an effort to validate that the job did not actually execute. When running these tests in isolation, they are passing. However, when run in the entire class, some of the preceding tests result in code that invokes TopicsWorker.getInstance() on a background thread, and this trips up the verifyNoMoreInteractions clause. Changing this to instead check that recordOnStartJob was never invoked works better. The other option is to put a sleep() between tests, which works but is not reliable and not desirable.

Bug: 289359982
Test: atest AdExtServicesServiceCoreUnitTests:EpochJobServiceTest
Change-Id: Ia3b3d90550ecaaf9b95702c97692b436bcb9c10d