Fix unnecessary stubbing bug in SdkSandboxManagerServiceUnitTest
As we use a fake callback instead of waiting for an actual result for
loadSdk method, unit tests can finish before SdkSandboxManagerService
does the network and foreground checks. If this happens, unnecessary
stubbing exception is thrown for disableForegroundCheck() and
disableNetworkPermissionChecks() methods. To solve the problem, 'lenient' strictness level should be used.

Bug: 289924609, Bug: 289924613, Bug: 289924755, Bug: 289923366
Test: atest SdkSandboxManagerServiceUnitTest
Change-Id: I14cc31869397b258443b66f7e2380d0057b99650