Skip tests that require usage of unsupported PackageManager APIs on R.
As noted in b/289120063, there are certain PackageManager APIs that are
not supported on R. While we are working on a solution (ag/23821381)
that provides an alternative solution to reading AdServices config from
the manfiest file without using PackageManager property related APIs,
there are still more tests that require explicit usage of unsupported
references to PacakageManager.property and
PackageManager::getResourcesForApplication(ApplicationInfo,
Configuration). These occur only in certain Topics related tests which
will continue to fail even after ag/23821381 is merged.

Since we are launching only Measurement API on R, we do not need a
workaround solution for these tests. Therefore, omitting these tests on
R.

Bug: b/289120063
Test: atest com.android.adservices.service.topics
Change-Id: Ida6e436dc1f04426f67a544947a0889894809002