DO NOT MERGE Delete all AdExtServices tests since the AdExtServices apk was removed from the ExtServices apex
This change needs to be in udc-dev, but it needs to NOT be in udc-mainline-prod and in main. Therefore, adding a Merged-In tag with the sha of a random commit that was merged into udc-mainline-prod. ag/24696255 was the latest commit, so using that.

Bug: 290781918
Test: atest
Change-Id: Icdfbb27aec8405a4f1a9549e170dc10af97f4904
Merged-In: I55f7c218e0d42b0272baa16c272a298639be1629