Move adId encryption from registration to matching
Details:
AdId needs to be persisted as original. Because XNA derived source's enrollment ID is
different from the original source's so the encrypted value will be
different. The derived sources are generated at the time of attribution,
so we'll pick the original adId and encrypt it with the trigger's
enrollment ID at the time of generating the reports.

Test: atest AdServicesServiceCoreUnitTests
Bug: 272356238
Change-Id: I9db7c05734787519b5f735891ae3870f3b4b7025