Fix flakyness for JS isolate being closed
[Symptom]
The test cases for verying the freeing of resources by closing JS Isolate
were flaky (60% + failure rate). This was due to the fact that there was
race condition between the start of execution on Isolate and the point
at which it was terminated.

[Cause]
Failures arised in the cases where the isolate was terminated before it
was even wrapped in a `Closeable`. This was due to incorrect wiring of a
half already computed future to the when->then of isolate'c mock

[Solution]
By wiring the doAnswer for mock, we ensured that latches for start of
execution are only triggered after the real mock-method call and not at the time
of wiring when -> then

Previous flaky run before fix : ab/I67700010177986556
Successfull runs after fix :ab/I01500010177860139

Bug: 240857630
Test: Fixed unit tests
Change-Id: Iaf76e76c6c8112a8bb58b3cc335162bc4d9ac795