Restructure OutcomeReceiver references in MeasurementManagerCtsTests.
Goal is to enable CtsAdExtServicesEndToEndTestMeasurement on Andorid R.
However, currently when the test suite is invoked on R it fails
immediately with "Cannot load android.os.OutcomeReceiver" errors. This
is because there are references to the OS receiver in the imports and in
method return types. Restructured these OS receiver references so that
the test suite at least executes on R without the OS receiver errors.
Also standardized naming of test methods to
testMethodName_setupConditions_expectedResult.

NOTE: Further investigation will need to be done to understand why tests are failing on R (not related to OS OutcomeReceiver references). These fixes will be made in a different CL.

Bug: 291633714
Test: atest CtsAdExtServicesEndToEndTestMeasurement
Change-Id: Iecfe67c0eae4cedd3fd3e3778b30cd88d0ac29d1