Parse DeviceConfig BroadcastReceiver canary set allowlist
A proto format is defined for a BroadcastReceiver canary allowlist.
This allowlist consists of a list of BroadcastReceiver intent action
strings.

This allowlist will be pushed via DeviceConfig in the form of a base64
encoded string which will need to be parsed and added to the existing
allowlist in SdkSandboxManagerService.

When an SDK tries to register a BroadcastReceiver, all intent filter
actions will be compared with the allowlist to check whether
registration should be allowed.

Bug: 291889486
Test: atest BroadcastRestrictionsHostTest
Change-Id: I8fb919a150d80e7c0aa811385bfd3774c852b7b6