Prefactoring Tests to allow to Mock Flags
We're working on ag/24060418 to enable a flag to guard DB schema change. This requires to introduce FlagFactory in DBHelper. It results in many measurement tests to fail due to Device_config access issue to read production flag in test. It causes AdServicesServiceCoreUnitTests:E2EDenoisedMockTest failure.

In this CL, I add a new constructor in DebugReportApi to get a
DbTestUtil.getMeasurementDbHelperForTest() in TestObjectProvider. We
should revert this CL in the future.

Bug: 291964983

Test: atest AdServicesServiceCoreUnitTests
Change-Id: I3783d7e8d906e6d4c9634f64d8d0a444619fa873