Move loadSdk System Server -> App logging to SandboxLatencyInfo
Latency logging for loadSdk's SYSTEM_SERVER_TO_APP stage is now logged
in logLatencies() method. For this, additional fields and methods are
added to the SandboxLatencyInfo class, and all needed timestamps are
saved directly to SandboxLatencyInfo instead of passing them as
parameters. LoadSdkReceiverProxy.logLatencyFromSystemServerToApp is
redundant now and so removed.

Bug: 287047664
Test: atest SdkSandboxFrameworkUnitTests
SdkSandboxManagerServiceUnitTests; statsd_testdrive 488 to check that atoms for the moved latency is reported

Change-Id: I52e8eb0507f43c12130c777996723168524d5b89