Move loadSdk System Server -> Sandbox -> System Server logging to SandboxLatencyInfo
Latency of these stages for loadSdk methos is now logged using SdkSandboxManagerService.logLatencies(). logLatencyMetricsForCallback in LoadSdkSession, which was used for this before, is still used for requestSurfacePackage and unloadSdk methods, and will be removed once latency logging for this methods is in SandboxLatencyInfo.

SandboxLatencyInfo is added as an argument of ISdkSandboxManager.loadSdk
method for testing purposes.

Bug: 287047664
Test: atest SdkSandboxFrameworkUnitTests
SdkSandboxManagerServiceUnitTests; statsd_testdrive 488 to check that
atoms for the moved latency is reported

Change-Id: I2673260b54830ed3333d11257796d440dac9c87f