Cobalt factory functions
Create Cobalt factory functions which return real implmentations of the
logger and periodic job.

Note, the period job DOES NOT upload data and results in data simply
being dropped. This is because the upload API had to be reverted for
M10 finalization and won't reland until that's done.

See b/294282666

Test: Factory function is trivial

Change-Id: I89bb781e50ee8274474c22c9333f0d9f4a8e7588