Parse DeviceConfig BroadcastReceiver allowlist
A proto format is defined for BroadcastReceiver allowlists in the form
of a map from effectiveTargetSdkVersion to a list of intent actions.
When restrictions are enforced, if all intent actions are present in the
allowlist, registration of the BroadcastReceiver will be permitted. If
the allowlist is not set at all, only protected broadcasts will be
allowed.

The allowlist will be pushed via DeviceConfig in the form of a base64
encoded string which is parsed by SdkSandboxManagerService.

Bug: 291889486
Test: atest BroadcastRestrictionsHostTest
Change-Id: Ib35839f7d85202348c53201a1f169e8d2e2152f4