Fix flaky EpochJobService tests.
EpochJobService tests are non-deterministic and flaky, especially on S.
Upon investigation, this flaky behavior is due to the fact that some
tests start an async job and immedialy assert for logging behavior.
While the job runs fast (usually < 1-2s), it's not guaranteed that the
logger has been invoked by the time we assert for logging. Added sleeps
where it makes sense so the test passing doesn't depend on timing of job
execution.

MUST_SLEEP

Bug: 293308741
Test: atest EpochJobServiceTest --iterations 15 (100% pass rate)
Change-Id: Ia64e77fe34128a280a7b8b98adc13fa51fdb94d9