Topics-specifc wrapper around Cobalt
The topics-specific wrapper translates topics to the event codes Cobalt
requires and handles setting the metric id. Note, the logger is designed
to accommodate upto 500 topics despite there there only being 446
currently. This is done so the mapping doesn't need to be updated if
more topics are added to the taxononmy. However it does assume topics
monotonically increase and can be mapped to Cobalt event codes simply by
subtracting 10000 from their value.

Test: atest TopicsCobaltLoggerTest

Change-Id: Ieb1abb6d18a4bc05ca667c5f06370d204fae5722