Update tests to run on device with small module pre-installed
The original condition of having com.android.adservices APEX module
pre-installed to run the test was too restrictive. The new update allows
the tests to run on device that already has small module pre-installed,
regardless of which signed version of the apex is loaded on device.

The TEST_MAPPING file has also been updated, but unlikely that any
automated tests configuration will pick it up since we don't have any
device configured with small-module pre-installed. That will be done
next.

Bug: 293842409
Test: atest SdkSandboxSmallModuleHostTest
Change-Id: I6041108a95f00f9d05af4fc963216c3a0cc71ca9