Fix crash due to STATUS_UNSET in FilterException
In rare cases when the filter classes throw an unexpected exception, the
resulting exception was getting wrapped in a FilterException and
translated into the status code STATUS_UNSET.  This status code was also
the default value in the FledgeErrorResponse, so that when the callback
for ad selection would fire its onError() call, an
IllegalArgumentException would crash the process.

The FilterException should instead use a valid status code when
unexpected errors are encountered.

Bug: 291086383
Test: atest AdServicesServiceCoreUnitTests
Change-Id: I2e8a2634278378df316048f0675a8a69b1a2a451