Fixes flaky SandboxedMeasurementManagerTest
Bug: 288333354
Issue reproduced locally by decreasing the max wait duraction for the
activity to start. Therefore, increasing the duration. Once the
activity starts, then the max duration will be ignored and proceed with
the test. The max duration was not getting respected and was treated as
min duration instead, this CL fixed that.

Test: atest CtsSandboxedMeasurementManagerTests --iterations=100
Change-Id: Ife97d75a29cc37a42890d7e904b273c5776f1dd3