Skip tests for SdkSandbox on Go devices by default
SdkSandbox will not be enabled on devices where AdServices is disabled.
So we are opting to use same device support logic for both suites.

Our tests are written with assumption that our module will be active.
Some devices can have a smaller module pre-loaded in them, in which
case these tests should be skipped. We didn't have proper assumptions
set in all of our tests. This cl fixes that.

In next cl, we will add CTS tests for devices with small module
pre-loaded and verify they run while everything else gets skipped.

Bug: 294504673
Test: atest --test-mapping
      packages/modules/AdServices/sdksandbox/TEST_MAPPING
Change-Id: Ic6e4c500418e3d1ad81c36132e99864fd3921d2a