Allow unsigned long range for dedup keys
Details: If a trigger is registered with aggregate or event report dedup
key value bigger than Long#MAX_VALUE, it is considered negative as a
signed long because most significant bit would be 1. At the time of retrieval,
UnsignedLong fails to parse it because it didn't expect a minus sign. If
the value is beyond UnsignedLong, the dedup key should be dropped, which
is already handled.
Also, handled existing negative values in the wild.

Bug: 294277936
Test: atest AdServicesServiceCoreUnitTests
Change-Id: I8e41b9ef8fc6c202c3bcd415b02c6d8be6994d4e