Prepend filenames with "adservices" on S-
AdServices code is included in the ExtServices APK on S-, and generates multiple files that are stored in the data directory during normal operation. After OTA to T, the AdService code is removed from the ExtServices APK, but these data files are left behind. We are adding a boot completed receiver to ExtServices to delete all adservices data after OTA. This change prepends all created files with the name "adservices" to make it easy to identify which files need to be removed. The underscore is for human readability; files that begin with "adservices" and have no underscore will still be deleted by the new boot completed receiver.

Test: atest AdServicesServiceCoreUnitTests, presubmit
Bug: 294912203
Change-Id: I5c45da9505b831b94fccd6852e52f77b432867a1