Simplify rules that depend on AbstractSupportedFeatureRule
AbstractSupportedFeatureRule was meant to make it easier to write
tests that automatically asserted that the code under test threw
an exception when ran in a device that didn't support the feature,
but in reality this logic made the tests that use the rule harder
to use / understand (as the unsupported check case is a corner case).

This CL removes the dependency on that superclass, which in a future
CL will be moved to a "experimental" package.

Bug: 295321663
Test: atest SdkSandboxManagerTest CtsAdServicesEndToEndTests CtsAdServicesHostTests

Change-Id: I52a7372632a2c4c5dd6c1e8b5b4652eb3e67ff9a