Fix missing UI CB test showmap metrics.
Showmap memory metrics are missing for UINotifLandingPage CB test. This
is because AdServices process is killed at the end of the test which
interferes with Showmap taking the snapshop of the process. We do not
need to kill the process at the end of the test because
UiNotificationLandingPageMicrobenchmark invokes the KillAppsRule, which
provides the same effect before each test. This test is now aligned with
the UiSettingsMainPage CB test for which we have appropriate Showmap
metrics.

Bug: 294605471
Test: ABTD - https://android-build.googleplex.com/builds/abtd/run/L48400000962579273. AdServices Showmap metrics appear under Artifacts.
Change-Id: Ibc459aa8220c8e0c04c3aacde2f9b549870b0961