Remove consent reset from test-level teardown
Instead of relying on a test-level teardown() method to reset flags and
consent, the module-level RunCommandTargetPreparer already handles
consent, enrollment, and API rate limit settings and resets for the
entire module.

This reset was bleeding into another test, causing it to flake when a
certain ordering of tests was run.

Test: atest CtsAdServicesDebuggableDeviceTestCases
Bug: 289670952
Change-Id: Ib50a1dea85429a5c39ba776edbaf8be8c613a46e