Log Error from IAPC Providers
When debugging the test issue, it throws ISE and unbinds from provider without any information. By adding local logging, it actually throws SecurityException due to newly added permission check. The SecurityException is not caught and crashes adservices process. Even worse, the error message is not logged and it returns "STATUS_INTERNAL_ERROR" instead. Eventually, it rethrows ISE defaultly due to "STATUS_INTERNAL_ERROR", which misleads the debugging.

Log the error message and add a TODO to handle the SecurityException properly

Bug: 294601334
Bug: 296928283
Test: non-logic change. Use pre-submit instead.
Change-Id: I8f91010ffe9f817a253f2f7dcad14831f5081af0