Improved error handling on AppImportanceFilter.assertCallerIsInForeground()
This method calls pm.getUidImportance(), which in turn throws a
SecurityException if the UID belongs to a different user (than the
caller). Currently, these exceptions are propagated to the caller,
which could crash the system server; this CL changes it to throw a
WrongCallingApplicationStateException instead (with is already
handled by AdServices components).

Test: atest AdExtServicesServiceCoreUnitTests:AppImportanceFilterTest#testSecurityExceptionTryingToRetrievePackageImportanceFromUid_throwsWrongCallingApplicationStateException
Bug: 295099457

Change-Id: I2fb89cc736e965851b9cde0fea74dcb0dbb1ec3d