Fixed AdServicesFlagsSetterRule to restore Shell permission on legacy mode.
CompatAdServicesTestUtils was refactored to use
AdServicesFlagsSetterRule, which wraps each call to DeviceConfig
undera a adopt / drop shell permissions. As such, tests that adopt
the shell permission on @Before, use CompatAdServicesTestUtils to
set flags, and then call DeviceConfig on @After would fail because
the permission would be gone at the latter.

This CL fixes that by re-adopting the permission when the rule is
used on "legacy mode".

Fixes: 296680559
Test: atest AdExtServicesEndToEndTests:CustomAudienceManagerTest \
            AdExtServicesEndToEndTests:TestAdSelectionManagerTest # on S

Change-Id: Iaea69645eaa3828576119cda476640ee5a425399