Remove consent reset from Measurement CTS teardown
Instead of relying on a test-level teardown() method to reset flags and
consent, the module-level RunCommandTargetPreparer already handles
consent, enrollment, and API rate limit settings and resets for the
entire module.

This reset was bleeding into another test, causing it to flake when a
certain ordering of tests was run.

Test: atest CtsAdServicesDebuggableDeviceTestCases
Bug: 289670952
Change-Id: I72628cd4dc4addc0d658106a6296ceef0bcce6e7