Fixed AdServicesSupportHelper.isLowRamDevice()
That method was meant to check if the device is Android Go (or more
specifically, if the device has a "lite" version of AdServices that
only defined the APIs but not the service implementation), but it
was just checking if the device had low memory. As such, it was
causing methods annotated with @RequiresLowRamDevice to fail
on non-Go devices that had low memory (like some watches).

Test: atest CtsAdServicesHostTests CtsAppSetIdEndToEndTest:com.android.adservices.tests.appsetid.AppSetIdManagerTest#testAppSetIdManager_whenDeviceNotSupported

Bug: 297408848
Bug: 298304610

Change-Id: I9eb4944541e46fb126b120a888311788376c852a