Handle incorrect WebView version internally
When calling selectAds() with an unsupported version of WebView, an
IllegalArgumentException was being thrown, but not caught or
appropriately handled.  This could lead to module rollback if the
exception was thrown frequently enough.

Replaced throwing in the ctors with failing in the execution so that
callbacks are appropriately called and the process won't crash.

Test: atest AdServicesServiceCoreUnitTests AdServicesJsEngineUnitTests
Bug: 298481006
Change-Id: Ib0c99d9cc5195ea1531d2af35e7f65e4b258241d