Allow AppManifestConfig to return enabled by default - part 2.
This CL allows AppManifestConfigHelper to return true to the
isAllowedApiXyzAccess() methods when the app manifest does not
contains the android.adservices.AD_SERVICES_CONFIG XML config.

Notice that this change only change that class and the affected
classes and unit tests - it's not effectly changing the behavior
at runtime yet, as that will be defined by a new flag.

Bug: 297585683

Test: atest \
  AdServicesServiceCoreUnitTests:AppManifestConfigParserTest \
  AdServicesServiceCoreUnitTests:AppManifestConfigHelperTest \
  AdServicesServiceCoreUnitTests:AndroidManifestConfigParserTest

Change-Id: I84be4bc31cf8e975321cdc9641f2f5fa28d32c34