Add CEL logging on aggregate key fetching errors
Bug: 298940688
Invalid parameter and json parser exceptions should not happen in
production as the aggregate key server is always the same. Adding logs
just in case. IOExceptions might happen, but should be relatively low.
Before calling the aggregate key server, checking to see if the user has
internet connectivity is performed and also the keys are cached for
about 1 week.

Test: atest AggregateEncryptionKeyFetcherTest

Change-Id: I591bbb403e0ff7dbf9a4e93f3430768d5ad96b9c