Manual App mediation cleaning
Change the name of AppOwnedSdk to MediateeSdk as it is used in app and
in sandbox as well.

MediateeSdk only need getMessage() for now, so it gets separated from
SdkApi, that makes SdkApi only implement the Api for the main
SampleSandboxedSdkProvider, and could be moved to its package, one
advantage to that, class R could be used (That is needed for using R to
load sandbox activity resources in further cls).

Test: Using manual app to show sd-to-sdk message in app and sandbox
Bug: 290333799
Change-Id: Ib6e75b4266c7dd4c0a3caaf15d31b604898b5aae