Refactor Cel to guard it behind the flag.
The flag guarding is present in AdServicesErrorLoggerImpl.java.
Before this flag, we are just passing the information and actual logging
logic is present in this class.
For example: onCreate method in TopicsService calls ->
ErrorLogUtil.e(int,int) calls ->
AdServicesErrorLoggerImpl.logError(int,int)

Also modified the ErrorLogUtil.logError signature to just include ppapi
and error_code and get method name and class name from the stack trace.
Thinking of CP this in M11.

Test: atest AdServicesServiceCoreUnitTests
Bug: 271147155
Bug: 299514253
Change-Id: I1695b7f65978741206b191fd790713af8a374993