Fix MeasurementService Race Condition.
Removing Flag and Null checks from MeasurementService::onBind. This will maintain the same functionality as before, without the redundancy or race condition that was crashing the service. As a bound service, an instance that tries to be bound while the kill switch is enabled will return null and be destroyed immediately as it has no bound applications.

Bug: 296464165
Test: atest AdServicesFrameworkUnitTests AdServicesServiceCoreUnitTests AdServicesManagerServiceTests AdServicesEndToEndTests
Change-Id: I7127da73957f9696cee2c5c31bb83f3dc8eea2cc