Execute ListenableFuture in the test to verify they throw ISE
For CustomAudienceManager, we have just updated the test.

For AdSelectionManager, we updated the test and reverted the changes
made in [1]. Implementation of AdSelectionManager expects
service to be null and when NPE occurs, it catches the exception and
returns it as an error through OutcomeReciever. This is preferrable
over throwing ISE and crashing the user.

[1] I1bef632a64df099a165e84fa955037a694bbee5c

Bug: 300875810
Test: atest AdSelectionTest#testGetAdSelectionService_lowRamDevice_throwsIllegalStateException
Test: atest CustomAudienceTest#testGetCustomAudienceService_lowRamDevice_throwsIllegalStateException
Change-Id: Ia8234f7d1b790d661bcb1ce569713fa939058a0e