Speed up AdServices host tests by not waiting blindly for boot complete
Currently, many of the host tests reboot the device and wait for 5 minutes, in order to give the boot-completed receiver enough time to have been invoked. However, in many cases, the receiver gets invoked pretty early on and the test is unnecessarily waiting. Helping speed up the test by only waiting until there is an indication of the boot-completed receiver running in the logs.

Fixed: 297111301
Test: atest CtsAdExtServicesHostTests
Change-Id: I5dc8310a73257a32d2e3c0f0f1f37f766fa04cb8