Add AIDL and system APIs for AdServicesExtDataStorageService.
AdServices have no rollback-exempted storage on Android R because neither AppSearch nor the system service exists on Android R.  As a result, we will be storing data in GMS Core (ProtoDataStore). In order to read and write data from the ProtoDataStore, this CL introduces new system APIs and interface for AdServicesExtDataStorageService which will be implemented in GMS core.

Bug: 303511619
Test: manual ; atest android.adservices.extdata
Change-Id: Ie614757faf3ed26a352129cad10ac40c9a2ce78c