Default allowlist for ContentProviders
Use a default allowlist for contentProviders from the sandbox process.
The default allowlist will be used when the allowlist for a specific SDK
target is not present.

This includes the case when the canary set of restrictions apply and the
canary set is empty, and so is the allowlist for the effective target
SDK version for the app which is making the request.

Bug: 274070295
Test: atest ContentProviderRestrictionsHostTest
Change-Id: Id7db19f590acfaee886d545875d32146f1cff080