Mark a couple of server auction cts tests as flaky
This test tries to make a network call to fetch the public keys. If the
test relies on an external network call, it is prone to flakiness due to
slow network.

One possible long term solution is to move this to root cts, start a
local server and fetch the key from there.

Test: not needed. marking a test as flaky
Change-Id: I44cadfd79d9ec820801325f4f72de3fba2fadf3f