Begin domain classes for the Cobalt registry
This CL is the first in a chain that replaces protobuf objects
with custom types inside the Cobalt client. This will allow the
client to use less memory, have types that codify invariants, e.g.
a single customer and projects, and more easily build abstractions
on top of the registry.

Later protobuf types will be moved to custom classes in follow up CLs,
where it makes sense.

Bug: 299492126

Test: atest AdServicesCobaltUnitTests
Test: atest AdServicesServiceCoreUnitTests

Change-Id: I63328561dd44c12212c7fdd93e9309c8ed8e8d02