Implement helper classes to interact with AdServicesExtDataStorageService.
Introducing manager and worker clases to read and write data from GMS
Core ProtoDataStore. For separation of tasks, the worker is responsible
for binding and interacting with GMS core asynchronously. The manager
class acts a wrapper over the worker class to await for callbacks with
set timeouts, and approrpiate debug logging. The manager class provides
user-friendly interfaces to access / update AdExt data and takes care of
error handling.

Bug: 307347006
Test: manual; atest com.android.adservices.service.extdata
Change-Id: I2b6c15f19de31fc781310b9e2d580a0f1a049a7e