Refactored how AbstractAdServicesFlagSetter cache commands.
Previously, it was using separate list for flags, system properites,
and commands; this CL merges that into a single list of commands
(which subclass for setting flags and system properties).

This change will guarantee that operations are executred in the
proper order (for example, to make sure that setSyncDisabledMode()
is always called first).

Bug: 294423183

Test: atest CtsAdServicesEndToEndTests:TopicsManagerTest CtsAdServicesHostTests # on T
Test: atest CtsAdExtServicesEndToEndTests:TopicsManagerTest CtsAdExtServicesHostTests # on R and S

Change-Id: Ic3bda654745de3b6b301038bd2a2eff5d2730420