Add latency logging for startSdkSandboxActivity
Time taken to complete the call is calculated in SdkSandboxManager, and
data is then sent to SdkSandboxManagerService to be logged to StatsD,
similarly to existing latency metrics.

Initially, it was planned to create a separate singleton logger that
would be used in framework, system server and sandbox processes and will
allow to avoid duplication of code, but this is currently blocked by
b/304340070: Statsd logger cannot be used in framework code. Until this
is resolved, latency will be logged directly from the places the method
is called, or from System server.

Design doc: go/rbs-activity-metrics
Bug: 300102108
Test: atest SdkSandboxFrameworkUnitTests
SdkSandboxManagerServiceUnitTests; statsd_testdrive -e 735 while
creating a fullscreen ad in the manual test app to ensure the atom is
logged

Change-Id: Iabb5c66ae15ab84e44f90eb08b8f4867610875f1