Move unloadSdk latency logging to SandboxLatencyInfo
As latency logging is now happenning in SdkSandboxManagerService, an
additional callback, IUnloadSdkCallback, is introduced to send latency
data from LoadSdkSession back to the service once unloading is done.
Existing sandbox callback is renamed to IUnloadSdkInSandboxCallback,
similarly to the existing loadSdk and requestSurfacePackage callbacks.

Bug: 287047664
Test: atest SdkSandboxFrameworkUnitTests
SdkSandboxManagerServiceUnitTests; statsd_testdrive 488 to check that
atoms for the moved latency is reported

Change-Id: Ibd59b45abbc5af966fae67ed809e977ca2999031