u18 notification fails with null on R
when source_of_truth = PPAPI_ONLY, u18 notification fails on R since
datastore returns null when it is used in if condition. this cl changes
datastore returns false instead of null

this failure only happens in testing. in production, we will use extdata
storage instead of PPAPI_ONLY. we are using PPAPI_ONLY in testing before
extdata is ready

Bug: 287245183
Test: atest + manual test u18 notificatio on R and verify bug is fixed
Change-Id: I9c6d66b8c024e1ce4b8061edc7386868f9bff283