Add latency logging for enforceAllowedToHostSandboxedActivity
Latency is calculated in SdkSandboxManagerLocal, and then sent to
SdkSandboxManagerService.logSandboxActivityEvent to be logged to StatsD.
In case any exception is thrown, the call is considered failed and
reported with 'FAILED_<exception name>' call result.

SDK uid is currently not logged for the method as it is called from
System Server -- this is tracked in b/305240130 and will be implemented
later along with startSdkSandboxActivity.

Design doc: go/rbs-activity-metrics
Bug: 300102108
Test: atest SdkSandboxManagerServiceUnitTests

Change-Id: I7d17966505be2bbdb6bfac64da850682d75d778b