Add total call sandbox latency logging
Total call latency is calculated in SandboxLatencyInfo after the call is
finished, and sent to StatsD along with latency for stages in
SdkSandboxManagerService.logLatencies(). Latency is calculated by
subtracting start and finish time of the call. Start time is the same
for all API method: timeAppCalledSystemServer. Finish time differs for
each API and is defined in SandboxLatencyInfo.getTotalCallFinishTime().

Bug: 288864263
Test: atest SdkSandboxFrameworkUnitTests
SdkSandboxManagerServiceUnitTests; statsd_testdrive 488

Change-Id: Ia6f86f8088cda98f4654170cdb4b5942bd57d367