Call isDebuggable prior to test start so the static variable initialization doesn't interfere with the test assertions
The current test has a 2-3% flakiness rate. The failure is because `SystemProperties.getInt` seems to be invoked on S even when it shouldn't. A potential cause for that could be that it was the first time that the `Build` class was ever being invoked so the `SystemProperties.getInt` was being invoked as part of the static initialization of that class. Therefore, ensuring that the `Build` class is statically initialized before the test starts tracking usages.

Bug: 311218889
Test: atest AdExtServicesServiceCoreUnitTests:BuildCompatUtilsTest
Change-Id: Ia22fecba91fb90028835789e4a1878a7abda9fc8