Add missing newline at the end of an API signature file
A previous change (ag/25287759) appears to have manually removed APIs,
rather than running `m update-api` and removed a blank line at the end
of the file that the `m checkapi` target expected to find. I think the
reason for that is the `m checkapi` target is not being run for the
`udc-mainline-prod` branch. cl/581880021 is intended to rectify that.

Test: presubmit
Bug: 309659592
Change-Id: I191a1607f33ef9614fb7f7478b8d5993dab383a4
Merged-In: I2814e349849e3eef6b17db5e1cee4fe6834faeb8